Changed correlateMessageService to own sink
For better control over where the messages are being sent
Also some refactoring
