+++ digiwf-spring-cloudstream-utils-starter/src/main/java/io/muenchendigital/digiwf/spring/cloudstream/utils/configuration/StreamingConfiguration.java
@@ -1,6 +1,6 @@
 package io.muenchendigital.digiwf.spring.cloudstream.utils.configuration;
 
-import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.RoutingCallback;
+import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.infrastructure.RoutingCallback;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
