+++ digiwf-spring-cloudstream-utils/src/main/java/io/muenchendigital/digiwf/spring/cloudstream/utils/api/streaming/infrastructure/MessageEmitter.java
@@ -1,5 +1,6 @@
-package io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming;
+package io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.infrastructure;
 
+import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.dto.CorrelateMessageDto;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.Message;
@@ -18,6 +19,16 @@ public class MessageEmitter {
 
     @Bean
     public Supplier<Flux<Message<Object>>> sendMessage(final Sinks.Many<Message<Object>> sink) {
+        return sink::asFlux;
+    }
+
+    @Bean
+    public Sinks.Many<Message<CorrelateMessageDto>> sendCorrelateMessageSink() {
+        return Sinks.many().unicast().onBackpressureBuffer();
+    }
+
+    @Bean
+    public Supplier<Flux<Message<CorrelateMessageDto>>> sendCorrelateMessage(final Sinks.Many<Message<CorrelateMessageDto>> sink) {
         return sink::asFlux;
     }
