+++ example/src/main/java/io/muenchendigital/digiwf/spring/cloudstream/api/controller/ExampleController.java
@@ -1,6 +1,6 @@
 package io.muenchendigital.digiwf.spring.cloudstream.api.controller;
 
-import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.PayloadSender;
+import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.service.PayloadSenderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -13,12 +13,12 @@ import javax.validation.constraints.NotNull;
 @RequiredArgsConstructor
 public class ExampleController {
 
-    private final PayloadSender genericPayloadSender;
+    private final PayloadSenderService genericPayloadSenderService;
 
     @PostMapping(value = "/sendMessage")
     public void sendMessage(@RequestBody @NotNull final String payload,
-                            @RequestParam(defaultValue = "processMessage") String type) {
-        this.genericPayloadSender.sendPayload(payload, type);
+                            @RequestParam(defaultValue = "processMessage") final String type) {
+        this.genericPayloadSenderService.sendPayload(payload, type);
     }
 
 }
