+++ example/src/main/resources/application.properties
@@ -6,5 +6,5 @@ spring.cloud.stream.kafka.binder.consumerProperties.value.deserializer=org.sprin
 spring.cloud.stream.kafka.binder.consumerProperties.auto.offset.reset=latest
 spring.cloud.stream.kafka.binder.configuration.security.protocol=SSL
 spring.kafka.consumer.properties.spring.json.trusted.packages=*
-spring.cloud.function.definition=functionRouter;sendMessage;
+spring.cloud.function.definition=functionRouter;sendMessage;sendCorrelateMessage;
 io.muenchendigital.digiwf.streaming.typeMappings.processMessage=processMessage
